requirements.txt
flask
openai
pandas
python-dotenv
pydantic
psycopg2-binary